requirements.txt
python-dotenv
uvicorn
fastapi
pymongo
aiogram==2.*
python-multipart